+++ server/Models/DTO/DiscountDTO.cs
@@ -4,7 +4,7 @@ public class DiscountGetDto
 {
     public int Id { get; set; }
     public string Type { get; set; } = "Percentage";
-    public decimal Value { get; set; }
+    public int Value { get; set; }
     public int Quantity;
     public string Code { get; set; } = string.Empty;
 }
@@ -20,7 +20,7 @@ public class DiscountCreateDto
 
     [Required(ErrorMessage = "Discount value is required")]
     [Range(0, double.MaxValue, ErrorMessage = "Discount value must be greater than or equal to 0")]
-    public decimal Value { get; set; }
+    public int Value { get; set; }
 
     [StringLength(50, ErrorMessage = "Discount code cannot exceed 50 characters")]
     public required string Code { get; set; }
@@ -41,7 +41,7 @@ public class DiscountPutDto
 
     [Required(ErrorMessage = "Discount value is required")]
     [Range(0, double.MaxValue, ErrorMessage = "Discount value must be greater than or equal to 0")]
-    public decimal Value { get; set; }
+    public int Value { get; set; }
 
     [StringLength(50, ErrorMessage = "Discount code cannot exceed 50 characters")]
     public required string Code { get; set; }
